// apps/sae/src/modules-features/admin/ModuleActivityParticipationTracking/ActivityEvidenceApproval/Interfaces/IEventViewModal.ts
export interface IEventViewModal {
  id: number;
  standardId: number;
  standardCode: string;
  standardName: string;
  code: string;
  name: string;
  hostName: string;
  reviewedName: string;
  completedName: string;
  addressName: string;
  quantity: number;
  maxPoint: number;
  minusPoint: number;
  facultyName: string;
  startDate: string;
  endDate: string;
  registered: number;
  checked: number;
  modifiedFullName: string;
  modifiedWhen: string;
}